moved oled defines to config.h and added ifndef to micro_oled driver
